(NewProject3D) Move the Has enum to world.ts
