+++ NewProject3D/components/com_camera.ts
@@ -1,7 +1,7 @@
 import {create} from "../../common/mat4.js";
 import {Mat4} from "../../common/math.js";
 import {Entity, Game} from "../game.js";
-import {Has} from "./com_index.js";
+import {Has} from "../world.js";
 
 export interface Camera {
     FovY: number;
