--- NewProject3D/components/com_index.ts
@@ -1,13 +1,0 @@
-const enum Component {
-    Camera,
-    Light,
-    Render,
-    Transform,
-}
-
-export const enum Has {
-    Camera = 1 << Component.Camera,
-    Light = 1 << Component.Light,
-    Render = 1 << Component.Render,
-    Transform = 1 << Component.Transform,
-}
