+++ NewProject3D/components/com_light.ts
@@ -1,6 +1,6 @@
 import {Vec3} from "../../common/math.js";
 import {Entity, Game} from "../game.js";
-import {Has} from "./com_index.js";
+import {Has} from "../world.js";
 
 export type Light = LightDirectional | LightPoint;
