+++ NewProject3D/components/com_render_diffuse.ts
@@ -3,7 +3,7 @@ import {Vec4} from "../../common/math.js";
 import {GL_ARRAY_BUFFER, GL_CW, GL_ELEMENT_ARRAY_BUFFER, GL_FLOAT} from "../../common/webgl.js";
 import {DiffuseLayout} from "../../materials/layout_diffuse.js";
 import {Entity, Game} from "../game.js";
-import {Has} from "./com_index.js";
+import {Has} from "../world.js";
 import {RenderKind} from "./com_render.js";
 
 export interface RenderDiffuse {
