+++ NewProject3D/systems/sys_camera.ts
@@ -1,6 +1,6 @@
 import {multiply, perspective} from "../../common/mat4.js";
-import {Has} from "../components/com_index.js";
 import {Entity, Game} from "../game.js";
+import {Has} from "../world.js";
 
 const QUERY = Has.Transform | Has.Camera;
