+++ NewProject3D/systems/sys_light.ts
@@ -1,9 +1,9 @@
 import {get_translation} from "../../common/mat4.js";
 import {Vec3} from "../../common/math.js";
 import {normalize} from "../../common/vec3.js";
-import {Has} from "../components/com_index.js";
 import {LightKind} from "../components/com_light.js";
 import {Entity, Game} from "../game.js";
+import {Has} from "../world.js";
 
 const QUERY = Has.Transform | Has.Light;
