+++ NewProject3D/systems/sys_transform.ts
@@ -1,8 +1,7 @@
 import {from_rotation_translation_scale, invert, multiply} from "../../common/mat4.js";
-import {Has} from "../components/com_index.js";
 import {Transform} from "../components/com_transform.js";
 import {Entity, Game} from "../game.js";
-import {World} from "../world.js";
+import {Has, World} from "../world.js";
 
 const QUERY = Has.Transform;
