+++ NewProject3D/world.ts
@@ -3,6 +3,20 @@ import {Light} from "./components/com_light.js";
 import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
+const enum Component {
+    Camera,
+    Light,
+    Render,
+    Transform,
+}
+
+export const enum Has {
+    Camera = 1 << Component.Camera,
+    Light = 1 << Component.Light,
+    Render = 1 << Component.Render,
+    Transform = 1 << Component.Transform,
+}
+
 export class World {
     // Component flags
     Mask: Array<number> = [];
